+++ src/Creature/CRTravelController.cpp
@@ -34,7 +34,7 @@ void CRTravelController::Init(){
 	}
 
 	goal = Vec3f(0,-10,0);
-	maxSpeed = 0.1;
+	maxSpeed = (float)0.1;
 }
 
 void CRTravelController::Step(){
@@ -57,10 +57,10 @@ void CRTravelController::Step(){
 		float distance = (goalPos - currPos).norm();
 
 		/**/
-		float limitStartDist = 2.0;
-		float limitSatDist   = 0.7;
-		float maxMax = 1.5;
-		float minMax = 0.1;
+		float limitStartDist = (float)2.0;
+		float limitSatDist   = (float)0.7;
+		float maxMax = (float)1.5;
+		float minMax = (float)0.1;
 		/*/
 		float limitStartDist = 2.0;
 		float limitSatDist   = 0.7;
@@ -98,8 +98,8 @@ void CRTravelController::Step(){
 		walkCtrl->SetRotationAngle(-asin(PTM::cross(currDir, dir)) * 3.0);
 		walkCtrl->SetSpeed((PTM::dot(currDir, dir)+1) * 1.0);
 		*/
-		walkCtrl->SetRotationAngle(-asin(PTM::cross(currDir, dir)) * 1.0); //walkCtrl->SetRotationAngle(-asin(PTM::cross(currDir, dir)) * 0.5);
-		walkCtrl->SetSpeed((PTM::dot(currDir, dir)+1) / 2.0 * maxSpeed);
+		walkCtrl->SetRotationAngle((float)(-asin(PTM::cross(currDir, dir)) * 1.0)); //walkCtrl->SetRotationAngle(-asin(PTM::cross(currDir, dir)) * 0.5);
+		walkCtrl->SetSpeed((float)((PTM::dot(currDir, dir)+1) / 2.0 * maxSpeed));
 	}
 }
